typescript for all our UIs
