+++ tgui/packages/tgui/interfaces/chompstation/CharacterDirectory/constants.ts
@@ -1,0 +1,18 @@
+export const getTagColor = (tag) => {
+  switch (tag) {
+    case 'Unset':
+      return 'label';
+    case 'Pred':
+      return 'red';
+    case 'Pred-Pref':
+      return 'orange';
+    case 'Prey':
+      return 'blue';
+    case 'Prey-Pref':
+      return 'green';
+    case 'Switch':
+      return 'yellow';
+    case 'Non-Vore':
+      return 'black';
+  }
+};
