+++ tgui/packages/tgui/interfaces/chompstation/ChemMaster/constants.ts
@@ -1,0 +1,8 @@
+export const transferAmounts: number[] = [1, 5, 10, 30, 60];
+export const bottleStyles: string[] = [
+  'bottle.png',
+  'small_bottle.png',
+  'wide_bottle.png',
+  'round_bottle.png',
+  'reagent_bottle.png',
+];
