+++ tgui/packages/tgui/interfaces/chompstation/ICAssembly.tsx
@@ -12,8 +12,22 @@ import {
 import { formatPower } from '../../format';
 import { Window } from '../../layouts';
 
+type Data = {
+  total_parts: number;
+  max_components: number;
+  total_complexity: number;
+  max_complexity: number;
+  battery_charge: number;
+  battery_max: number;
+  net_power: number;
+  unremovable_circuits: circuit[];
+  removable_circuits: circuit[];
+};
+
+type circuit = { name: string; ref: string };
+
 export const ICAssembly = (props) => {
-  const { act, data } = useBackend();
+  const { act, data } = useBackend<Data>();
 
   const {
     total_parts,
@@ -28,7 +42,7 @@ export const ICAssembly = (props) => {
   } = data;
 
   return (
-    <Window width={600} height={380} resizable>
+    <Window width={600} height={380}>
       <Window.Content scrollable>
         <Section
           title="Status"
