+++ tgui/packages/tgui/interfaces/chompstation/TguiFeedback.tsx
@@ -20,8 +20,8 @@ type Data = {
 export const TguiFeedback = (props) => {
   const { act, data } = useBackend<Data>();
   const { open_windows, selected_window } = data;
-  const [rating, setRating] = useState('');
-  const [comment, setComment] = useState('');
+  const [rating, setRating] = useState<string>('');
+  const [comment, setComment] = useState<string>('');
   return (
     <Window width={550} height={460}>
       <Flex>
