+++ tgui/packages/tgui/interfaces/chompstation/Ticket.tsx
@@ -1,8 +1,8 @@
 /* eslint react/no-danger: "off" */
+import { KEY } from 'common/keys';
 import { round, toFixed } from 'common/math';
 import { useState } from 'react';
 
-import { KEY_ENTER } from '../../../common/keycodes';
 import { useBackend } from '../../backend';
 import {
   Box,
@@ -54,12 +54,15 @@ type Data = {
   log: string[];
 };
 
+window.addEventListener('keydown', (event) => {
+  console.log(event);
+});
+
 export const Ticket = (props) => {
   const { act, data } = useBackend<Data>();
   const [ticketChat, setTicketChat] = useState('');
   const {
     id,
-    title,
     name,
     ticket_ref,
     state,
@@ -134,12 +137,9 @@ export const Ticket = (props) => {
                     fluid
                     placeholder="Enter a message..."
                     value={ticketChat}
-                    onInput={(e, value) => setTicketChat(value)}
-                    onKeyDown={(event) => {
-                      const keyCode = window.event
-                        ? event.which
-                        : event.keyCode;
-                      if (keyCode === KEY_ENTER) {
+                    onInput={(e, value: string) => setTicketChat(value)}
+                    onKeyDown={(e) => {
+                      if (KEY.Enter === e.key) {
                         act('send_msg', {
                           msg: ticketChat,
                           ticket_ref: ticket_ref,
