+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreBellySelectionAndCustomization.tsx
@@ -1,3 +1,5 @@
+import { BooleanLike } from 'common/react';
+
 import { useBackend } from '../../../backend';
 import { Box, Divider, Flex, Icon, Section, Tabs } from '../../../components';
 import { digestModeToColor } from './constants';
@@ -7,10 +9,12 @@ import { VoreSelectedBelly } from './VoreSelectedBelly';
 export const VoreBellySelectionAndCustomization = (props: {
   our_bellies: bellyData[];
   selected: selectedData;
+  show_pictures: BooleanLike;
+  icon_overflow: BooleanLike;
 }) => {
   const { act } = useBackend();
 
-  const { our_bellies, selected } = props;
+  const { our_bellies, selected, show_pictures, icon_overflow } = props;
 
   return (
     <Flex>
@@ -60,7 +64,11 @@ export const VoreBellySelectionAndCustomization = (props: {
       <Flex.Item grow>
         {selected && (
           <Section title={selected.belly_name}>
-            <VoreSelectedBelly belly={selected} />
+            <VoreSelectedBelly
+              belly={selected}
+              show_pictures={show_pictures}
+              icon_overflow={icon_overflow}
+            />
           </Section>
         )}
       </Flex.Item>
