+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBelly.tsx
@@ -1,3 +1,4 @@
+import { BooleanLike } from 'common/react';
 import { useState } from 'react';
 
 import { Tabs } from '../../../components';
@@ -14,8 +15,12 @@ import { VoreSelectedBellyVisuals } from './VoreSelectedBellyVisuals';
 /**
  * Subtemplate of VoreBellySelectionAndCustomization
  */
-export const VoreSelectedBelly = (props: { belly: selectedData }) => {
-  const { belly } = props;
+export const VoreSelectedBelly = (props: {
+  belly: selectedData;
+  show_pictures: BooleanLike;
+  icon_overflow: BooleanLike;
+}) => {
+  const { belly, show_pictures, icon_overflow } = props;
   const { contents } = belly;
 
   const [tabIndex, setTabIndex] = useState(0);
@@ -28,7 +33,14 @@ export const VoreSelectedBelly = (props: { belly: selectedData }) => {
   tabs[3] = <VoreSelectedBellySounds belly={belly} />;
   tabs[4] = <VoreSelectedBellyVisuals belly={belly} />;
   tabs[5] = <VoreSelectedBellyInteractions belly={belly} />;
-  tabs[6] = <VoreContentsPanel outside contents={contents} />;
+  tabs[6] = (
+    <VoreContentsPanel
+      outside
+      contents={contents}
+      show_pictures={show_pictures}
+      icon_overflow={icon_overflow}
+    />
+  );
   tabs[7] = <VoreSelectedBellyLiquidOptions belly={belly} />;
   tabs[8] = <VoreSelectedBellyLiquidMessages belly={belly} />;
