+++ tgui/packages/tgui/interfaces/chompstation/VorePanelExport/constants.ts
@@ -1,0 +1,62 @@
+const ModeSpan = {
+  Hold: '<span class="badge text-bg-secondary">Hold</span>',
+  Digest: '<span class="badge text-bg-danger">Digest</span>',
+  Absorb: '<span class="badge text-bg-light">Absorb</span>',
+  Drain: '<span class="badge text-bg-warning">Drain</span>',
+  Selective: '<span class="badge text-bg-warning">Selective</span>',
+  Unabsorb: '<span class="badge text-bg-light">Unabsorb</span>',
+  Heal: '<span class="badge text-bg-success">Heal</span>',
+  Shrink: '<span class="badge text-bg-info">Shrink</span>',
+  Grow: '<span class="badge text-bg-info">Grow</span>',
+  'Size Steal': '<span class="badge text-bg-info">Size Steal</span>',
+  'Encase In Egg': '<span class="badge text-bg-primary">Encase In Egg</span>',
+};
+
+const ItemModeSpan = {
+  Hold: '<span class="badge text-bg-secondary">Item: Hold</span>',
+  'Digest (Food Only)':
+    '<span class="badge text-bg-danger">Item: Digest (Food Only)</span>',
+  Digest: '<span class="badge text-bg-danger">Item: Digest</span>',
+  'Digest (Dispersed Damage)':
+    '<span class="badge text-bg-danger">Item: Digest (Dispersed Damage)</span>',
+};
+
+const AddonIcon = {
+  Numbing: '',
+  Stripping: '',
+  'Leave Remains': '',
+  Muffles: 'bi-volume-mute',
+  'Affect Worn Items': '',
+  'Jams Sensors': 'bi-wifi-off',
+  'Complete Absorb': '',
+};
+
+const ReagentAddonIcon = {
+  'Produce Liquids': '',
+  'Digestion Liquids': '',
+  'Absorption Liquids': '',
+  'Draining Liquids': '',
+};
+
+const AutotransferFlagIcon = {
+  Creatures: '',
+  Absorbed: '',
+  Carbon: '',
+  Silicon: '',
+  Mobs: '',
+  Animals: '',
+  Mice: '',
+  Dead: '',
+  'Digestable Creatures': '',
+  'Absorbable Creatures': '',
+  'Full Health': '',
+  Items: '',
+  Trash: '',
+  Eggs: '',
+  Remains: '',
+  'Indigestible Items': '',
+  'Recyclable Items': '',
+  Ores: '',
+  'Clothes and Bags': '',
+  Food: '',
+};
